# server/api.yaml
openapi: 3.0.3
info:
  title: Calorimeter API
  version: "1.0.0"
  description: "An API for managing calorimeter records"
paths:
  /store:
    post:
      operationId: bulkStoreRecords
      summary: Bulk store records
      description: >
        Store one or more records. Each record includes a user_id, a list of numeric values,
        and an optional description. Returns the UUIDs of the saved records.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreRequest'
      responses:
        '200':
          description: Successfully stored records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreResponse'
  /records:
    post:
      operationId: retrieveRecords
      summary: Retrieve records for a user
      description: >
        Retrieve records for a given user. The request body must include the user_id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordsRequest'
      responses:
        '200':
          description: A list of records for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecordOutput'
  /update_user:
    post:
      operationId: updateUser
      summary: Update user information
      description: >
        Update user information for a given user. The request body must include the user_id 
        and optional fields for surname and name.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
  /user_info:
    post:
      operationId: getUserInfo
      summary: Retrieve user information
      description: >
        Get the user information for the specified user. The request body must include the user_id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIdRequest'
      responses:
        '200':
          description: Successfully retrieved user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
components:
  schemas:
    StoreRequestItem:
      type: object
      properties:
        value:
          type: integer
          description: Numeric value associated with the record
        description:
          type: string
          description: Optional description for the record
      required:
        - user_id
        - value
    StoreRequest:
      type: object
      properties:
        user_id:
          type: integer
          description: ID of the user creating the record
        items:
          type: array
          items:
            $ref: '#/components/schemas/StoreRequestItem'
      required:
        - user_id
        - items
    StoreResponse:
      type: object
      properties:
        uuids:
          type: array
          items:
            type: string
          description: UUIDs of the saved records
      required:
        - uuids

    RecordsRequest:
      type: object
      properties:
        user_id:
          type: integer
          description: ID of the user
      required:
        - user_id
    RecordOutput:
      type: object
      properties:
        uuid:
          type: string
          description: Unique identifier of the record
        user_id:
          type: integer
          description: ID of the user associated with the record

    UserUpdateRequest:
      type: object
      properties:
        user_id:
          type: integer
          description: ID of the user to update
        surname:
          type: string
          description: Optional surname of the user
        name:
          type: string
          description: Optional first name of the user
      required:
        - user_id
    UserInfo:
      type: object
      properties:
        user_id:
          type: integer
          description: User's ID
        surname:
          type: string
          description: User's surname
        name:
          type: string
          description: User's first name

    UserIdRequest:
      type: object
      properties:
        user_id:
          type: integer
          description: ID of the user
      required:
        - user_id